+++ substratevm/src/com.oracle.svm.core/src/com/oracle/svm/core/jdk/VarHandleFeature.java
@@ -41,18 +41,22 @@ import org.graalvm.nativeimage.ImageSingletons;
 
 import com.oracle.svm.core.StaticFieldsSupport;
 import com.oracle.svm.core.annotate.Alias;
-import com.oracle.svm.core.annotate.InjectAccessors;
 import com.oracle.svm.core.annotate.RecomputeFieldValue;
 import com.oracle.svm.core.annotate.RecomputeFieldValue.Kind;
 import com.oracle.svm.core.annotate.TargetClass;
 import com.oracle.svm.core.feature.AutomaticallyRegisteredFeature;
 import com.oracle.svm.core.feature.InternalFeature;
 import com.oracle.svm.core.fieldvaluetransformer.FieldValueTransformerWithAvailability;
+import com.oracle.svm.core.graal.nodes.FieldOffsetNode;
 import com.oracle.svm.core.reflect.target.ReflectionSubstitutionSupport;
 import com.oracle.svm.core.util.VMError;
 import com.oracle.svm.util.ReflectionUtil;
 
+import jdk.graal.compiler.nodes.ValueNode;
+import jdk.graal.compiler.nodes.spi.CoreProviders;
 import jdk.internal.misc.Unsafe;
+import jdk.vm.ci.meta.JavaConstant;
+import jdk.vm.ci.meta.JavaKind;
 
 /**
  * This file contains most of the code necessary for supporting VarHandle (and DirectMethodHandle
@@ -84,10 +88,7 @@ import jdk.internal.misc.Unsafe;
  * same custom field value recomputation handler: {@link VarHandleFieldOffsetComputer}.
  *
  * For static fields, also the base of the Unsafe access needs to be changed to the static field
- * holder arrays defined in {@link StaticFieldsSupport}. We cannot do a recomputation to the actual
- * arrays because the arrays are only available after static analysis. So we inject accessor methods
- * instead that read our holder fields: {@link VarHandleFieldStaticBasePrimitiveAccessor} and
- * {@link VarHandleFieldStaticBaseObjectAccessor}.
+ * holder arrays defined in {@link StaticFieldsSupport}: {@link VarHandleStaticBaseComputer}.
  *
  * VarHandle access to arrays is the simplest case: we only need field value recomputations for the
  * array base offset and array index shift.
@@ -307,7 +308,14 @@ class VarHandleInfo {
     }
 }
 
-class VarHandleFieldOffsetIntComputer implements FieldValueTransformerWithAvailability {
+abstract class VarHandleFieldOffsetComputer implements FieldValueTransformerWithAvailability {
+
+    private final JavaKind kind;
+
+    VarHandleFieldOffsetComputer(JavaKind kind) {
+        this.kind = kind;
+    }
+
     @Override
     public ValueAvailability valueAvailability() {
         return ValueAvailability.AfterAnalysis;
@@ -320,37 +328,60 @@ class VarHandleFieldOffsetIntComputer implements FieldValueTransformerWithAvaila
         if (offset <= 0) {
             throw VMError.shouldNotReachHere("Field is not marked as unsafe accessed: " + field);
         }
-        return offset;
+
+        switch (kind) {
+            case Int:
+                return Integer.valueOf(offset);
+            case Long:
+                return Long.valueOf(offset);
+            default:
+                throw VMError.shouldNotReachHere("Invalid kind: " + kind);
+        }
+    }
+
+    @Override
+    public ValueNode intrinsify(CoreProviders providers, JavaConstant receiver) {
+        Object varHandle = providers.getSnippetReflection().asObject(Object.class, receiver);
+        if (varHandle != null) {
+            Field field = ImageSingletons.lookup(VarHandleFeature.class).findVarHandleField(varHandle);
+            return FieldOffsetNode.create(kind, providers.getMetaAccess().lookupJavaField(field));
+        }
+        return null;
     }
 }
 
-class VarHandleFieldOffsetComputer extends VarHandleFieldOffsetIntComputer {
+class VarHandleFieldOffsetAsIntComputer extends VarHandleFieldOffsetComputer {
+    VarHandleFieldOffsetAsIntComputer() {
+        super(JavaKind.Int);
+    }
+}
+
+class VarHandleFieldOffsetAsLongComputer extends VarHandleFieldOffsetComputer {
+    VarHandleFieldOffsetAsLongComputer() {
+        super(JavaKind.Long);
+    }
+}
+
+class VarHandleStaticBaseComputer implements FieldValueTransformerWithAvailability {
+    @Override
+    public ValueAvailability valueAvailability() {
+        return ValueAvailability.AfterAnalysis;
+    }
+
     @Override
     public Object transform(Object receiver, Object originalValue) {
-        Object offset = super.transform(receiver, originalValue);
-        return Long.valueOf((Integer) offset);
-    }
-}
-
-class VarHandleFieldStaticBasePrimitiveAccessor {
-    static Object get(@SuppressWarnings("unused") Object varHandle) {
-        return StaticFieldsSupport.getStaticPrimitiveFields();
+        Field field = ImageSingletons.lookup(VarHandleFeature.class).findVarHandleField(receiver);
+        return field.getType().isPrimitive() ? StaticFieldsSupport.getStaticPrimitiveFields() : StaticFieldsSupport.getStaticObjectFields();
     }
 
-    @SuppressWarnings("unused")
-    static void set(Object varHandle, Object value) {
-        assert value == StaticFieldsSupport.getStaticPrimitiveFields();
-    }
-}
-
-class VarHandleFieldStaticBaseObjectAccessor {
-    static Object get(@SuppressWarnings("unused") Object varHandle) {
-        return StaticFieldsSupport.getStaticObjectFields();
-    }
-
-    @SuppressWarnings("unused")
-    static void set(Object varHandle, Object value) {
-        assert value == StaticFieldsSupport.getStaticObjectFields();
+    @Override
+    public ValueNode intrinsify(CoreProviders providers, JavaConstant receiver) {
+        Object varHandle = providers.getSnippetReflection().asObject(Object.class, receiver);
+        if (varHandle != null) {
+            Field field = ImageSingletons.lookup(VarHandleFeature.class).findVarHandleField(varHandle);
+            return StaticFieldsSupport.createStaticFieldBaseNode(field.getType().isPrimitive());
+        }
+        return null;
     }
 }
 
@@ -448,55 +479,55 @@ final class Target_java_lang_invoke_VarHandleReferences_Array {
 
 @TargetClass(className = "java.lang.invoke.VarHandleBooleans", innerClass = "FieldInstanceReadOnly")
 final class Target_java_lang_invoke_VarHandleBooleans_FieldInstanceReadOnly {
-    @Alias @RecomputeFieldValue(kind = Kind.Custom, declClass = VarHandleFieldOffsetComputer.class) //
+    @Alias @RecomputeFieldValue(kind = Kind.Custom, declClass = VarHandleFieldOffsetAsLongComputer.class) //
     long fieldOffset;
 }
 
 @TargetClass(className = "java.lang.invoke.VarHandleBytes", innerClass = "FieldInstanceReadOnly")
 final class Target_java_lang_invoke_VarHandleBytes_FieldInstanceReadOnly {
-    @Alias @RecomputeFieldValue(kind = Kind.Custom, declClass = VarHandleFieldOffsetComputer.class) //
+    @Alias @RecomputeFieldValue(kind = Kind.Custom, declClass = VarHandleFieldOffsetAsLongComputer.class) //
     long fieldOffset;
 }
 
 @TargetClass(className = "java.lang.invoke.VarHandleChars", innerClass = "FieldInstanceReadOnly")
 final class Target_java_lang_invoke_VarHandleChars_FieldInstanceReadOnly {
-    @Alias @RecomputeFieldValue(kind = Kind.Custom, declClass = VarHandleFieldOffsetComputer.class) //
+    @Alias @RecomputeFieldValue(kind = Kind.Custom, declClass = VarHandleFieldOffsetAsLongComputer.class) //
     long fieldOffset;
 }
 
 @TargetClass(className = "java.lang.invoke.VarHandleDoubles", innerClass = "FieldInstanceReadOnly")
 final class Target_java_lang_invoke_VarHandleDoubles_FieldInstanceReadOnly {
-    @Alias @RecomputeFieldValue(kind = Kind.Custom, declClass = VarHandleFieldOffsetComputer.class) //
+    @Alias @RecomputeFieldValue(kind = Kind.Custom, declClass = VarHandleFieldOffsetAsLongComputer.class) //
     long fieldOffset;
 }
 
 @TargetClass(className = "java.lang.invoke.VarHandleFloats", innerClass = "FieldInstanceReadOnly")
 final class Target_java_lang_invoke_VarHandleFloats_FieldInstanceReadOnly {
-    @Alias @RecomputeFieldValue(kind = Kind.Custom, declClass = VarHandleFieldOffsetComputer.class) //
+    @Alias @RecomputeFieldValue(kind = Kind.Custom, declClass = VarHandleFieldOffsetAsLongComputer.class) //
     long fieldOffset;
 }
 
 @TargetClass(className = "java.lang.invoke.VarHandleInts", innerClass = "FieldInstanceReadOnly")
 final class Target_java_lang_invoke_VarHandleInts_FieldInstanceReadOnly {
-    @Alias @RecomputeFieldValue(kind = Kind.Custom, declClass = VarHandleFieldOffsetComputer.class) //
+    @Alias @RecomputeFieldValue(kind = Kind.Custom, declClass = VarHandleFieldOffsetAsLongComputer.class) //
     long fieldOffset;
 }
 
 @TargetClass(className = "java.lang.invoke.VarHandleLongs", innerClass = "FieldInstanceReadOnly")
 final class Target_java_lang_invoke_VarHandleLongs_FieldInstanceReadOnly {
-    @Alias @RecomputeFieldValue(kind = Kind.Custom, declClass = VarHandleFieldOffsetComputer.class) //
+    @Alias @RecomputeFieldValue(kind = Kind.Custom, declClass = VarHandleFieldOffsetAsLongComputer.class) //
     long fieldOffset;
 }
 
 @TargetClass(className = "java.lang.invoke.VarHandleShorts", innerClass = "FieldInstanceReadOnly")
 final class Target_java_lang_invoke_VarHandleShorts_FieldInstanceReadOnly {
-    @Alias @RecomputeFieldValue(kind = Kind.Custom, declClass = VarHandleFieldOffsetComputer.class) //
+    @Alias @RecomputeFieldValue(kind = Kind.Custom, declClass = VarHandleFieldOffsetAsLongComputer.class) //
     long fieldOffset;
 }
 
 @TargetClass(className = "java.lang.invoke.VarHandleReferences", innerClass = "FieldInstanceReadOnly")
 final class Target_java_lang_invoke_VarHandleReferences_FieldInstanceReadOnly {
-    @Alias @RecomputeFieldValue(kind = Kind.Custom, declClass = VarHandleFieldOffsetComputer.class) //
+    @Alias @RecomputeFieldValue(kind = Kind.Custom, declClass = VarHandleFieldOffsetAsLongComputer.class) //
     long fieldOffset;
 }
 
@@ -509,73 +540,73 @@ final class Target_java_lang_invoke_VarHandleReferences_FieldInstanceReadOnly {
 
 @TargetClass(className = "java.lang.invoke.VarHandleBooleans", innerClass = "FieldStaticReadOnly")
 final class Target_java_lang_invoke_VarHandleBooleans_FieldStaticReadOnly {
-    @Alias @InjectAccessors(VarHandleFieldStaticBasePrimitiveAccessor.class) //
+    @Alias @RecomputeFieldValue(kind = Kind.Custom, declClass = VarHandleStaticBaseComputer.class) //
     Object base;
-    @Alias @RecomputeFieldValue(kind = Kind.Custom, declClass = VarHandleFieldOffsetComputer.class) //
+    @Alias @RecomputeFieldValue(kind = Kind.Custom, declClass = VarHandleFieldOffsetAsLongComputer.class) //
     long fieldOffset;
 }
 
 @TargetClass(className = "java.lang.invoke.VarHandleBytes", innerClass = "FieldStaticReadOnly")
 final class Target_java_lang_invoke_VarHandleBytes_FieldStaticReadOnly {
-    @Alias @InjectAccessors(VarHandleFieldStaticBasePrimitiveAccessor.class) //
+    @Alias @RecomputeFieldValue(kind = Kind.Custom, declClass = VarHandleStaticBaseComputer.class) //
     Object base;
-    @Alias @RecomputeFieldValue(kind = Kind.Custom, declClass = VarHandleFieldOffsetComputer.class) //
+    @Alias @RecomputeFieldValue(kind = Kind.Custom, declClass = VarHandleFieldOffsetAsLongComputer.class) //
     long fieldOffset;
 }
 
 @TargetClass(className = "java.lang.invoke.VarHandleChars", innerClass = "FieldStaticReadOnly")
 final class Target_java_lang_invoke_VarHandleChars_FieldStaticReadOnly {
-    @Alias @InjectAccessors(VarHandleFieldStaticBasePrimitiveAccessor.class) //
+    @Alias @RecomputeFieldValue(kind = Kind.Custom, declClass = VarHandleStaticBaseComputer.class) //
     Object base;
-    @Alias @RecomputeFieldValue(kind = Kind.Custom, declClass = VarHandleFieldOffsetComputer.class) //
+    @Alias @RecomputeFieldValue(kind = Kind.Custom, declClass = VarHandleFieldOffsetAsLongComputer.class) //
     long fieldOffset;
 }
 
 @TargetClass(className = "java.lang.invoke.VarHandleDoubles", innerClass = "FieldStaticReadOnly")
 final class Target_java_lang_invoke_VarHandleDoubles_FieldStaticReadOnly {
-    @Alias @InjectAccessors(VarHandleFieldStaticBasePrimitiveAccessor.class) //
+    @Alias @RecomputeFieldValue(kind = Kind.Custom, declClass = VarHandleStaticBaseComputer.class) //
     Object base;
-    @Alias @RecomputeFieldValue(kind = Kind.Custom, declClass = VarHandleFieldOffsetComputer.class) //
+    @Alias @RecomputeFieldValue(kind = Kind.Custom, declClass = VarHandleFieldOffsetAsLongComputer.class) //
     long fieldOffset;
 }
 
 @TargetClass(className = "java.lang.invoke.VarHandleFloats", innerClass = "FieldStaticReadOnly")
 final class Target_java_lang_invoke_VarHandleFloats_FieldStaticReadOnly {
-    @Alias @InjectAccessors(VarHandleFieldStaticBasePrimitiveAccessor.class) //
+    @Alias @RecomputeFieldValue(kind = Kind.Custom, declClass = VarHandleStaticBaseComputer.class) //
     Object base;
-    @Alias @RecomputeFieldValue(kind = Kind.Custom, declClass = VarHandleFieldOffsetComputer.class) //
+    @Alias @RecomputeFieldValue(kind = Kind.Custom, declClass = VarHandleFieldOffsetAsLongComputer.class) //
     long fieldOffset;
 }
 
 @TargetClass(className = "java.lang.invoke.VarHandleInts", innerClass = "FieldStaticReadOnly")
 final class Target_java_lang_invoke_VarHandleInts_FieldStaticReadOnly {
-    @Alias @InjectAccessors(VarHandleFieldStaticBasePrimitiveAccessor.class) //
+    @Alias @RecomputeFieldValue(kind = Kind.Custom, declClass = VarHandleStaticBaseComputer.class) //
     Object base;
-    @Alias @RecomputeFieldValue(kind = Kind.Custom, declClass = VarHandleFieldOffsetComputer.class) //
+    @Alias @RecomputeFieldValue(kind = Kind.Custom, declClass = VarHandleFieldOffsetAsLongComputer.class) //
     long fieldOffset;
 }
 
 @TargetClass(className = "java.lang.invoke.VarHandleLongs", innerClass = "FieldStaticReadOnly")
 final class Target_java_lang_invoke_VarHandleLongs_FieldStaticReadOnly {
-    @Alias @InjectAccessors(VarHandleFieldStaticBasePrimitiveAccessor.class) //
+    @Alias @RecomputeFieldValue(kind = Kind.Custom, declClass = VarHandleStaticBaseComputer.class) //
     Object base;
-    @Alias @RecomputeFieldValue(kind = Kind.Custom, declClass = VarHandleFieldOffsetComputer.class) //
+    @Alias @RecomputeFieldValue(kind = Kind.Custom, declClass = VarHandleFieldOffsetAsLongComputer.class) //
     long fieldOffset;
 }
 
 @TargetClass(className = "java.lang.invoke.VarHandleShorts", innerClass = "FieldStaticReadOnly")
 final class Target_java_lang_invoke_VarHandleShorts_FieldStaticReadOnly {
-    @Alias @InjectAccessors(VarHandleFieldStaticBasePrimitiveAccessor.class) //
+    @Alias @RecomputeFieldValue(kind = Kind.Custom, declClass = VarHandleStaticBaseComputer.class) //
     Object base;
-    @Alias @RecomputeFieldValue(kind = Kind.Custom, declClass = VarHandleFieldOffsetComputer.class) //
+    @Alias @RecomputeFieldValue(kind = Kind.Custom, declClass = VarHandleFieldOffsetAsLongComputer.class) //
     long fieldOffset;
 }
 
 @TargetClass(className = "java.lang.invoke.VarHandleReferences", innerClass = "FieldStaticReadOnly")
 final class Target_java_lang_invoke_VarHandleReferences_FieldStaticReadOnly {
-    @Alias @InjectAccessors(VarHandleFieldStaticBaseObjectAccessor.class) //
+    @Alias @RecomputeFieldValue(kind = Kind.Custom, declClass = VarHandleStaticBaseComputer.class) //
     Object base;
-    @Alias @RecomputeFieldValue(kind = Kind.Custom, declClass = VarHandleFieldOffsetComputer.class) //
+    @Alias @RecomputeFieldValue(kind = Kind.Custom, declClass = VarHandleFieldOffsetAsLongComputer.class) //
     long fieldOffset;
 }
 
@@ -586,32 +617,14 @@ final class Target_java_lang_invoke_VarHandleReferences_FieldStaticReadOnly {
 
 @TargetClass(className = "java.lang.invoke.DirectMethodHandle", innerClass = "Accessor")
 final class Target_java_lang_invoke_DirectMethodHandle_Accessor {
-    @Alias @RecomputeFieldValue(kind = RecomputeFieldValue.Kind.Custom, declClass = VarHandleFieldOffsetIntComputer.class) //
+    @Alias @RecomputeFieldValue(kind = RecomputeFieldValue.Kind.Custom, declClass = VarHandleFieldOffsetAsIntComputer.class) //
     int fieldOffset;
 }
 
 @TargetClass(className = "java.lang.invoke.DirectMethodHandle", innerClass = "StaticAccessor")
 final class Target_java_lang_invoke_DirectMethodHandle_StaticAccessor {
-    @Alias //
-    Class<?> fieldType;
-    @Alias @RecomputeFieldValue(kind = RecomputeFieldValue.Kind.Custom, declClass = StaticAccessorFieldStaticBaseComputer.class) //
+    @Alias @RecomputeFieldValue(kind = RecomputeFieldValue.Kind.Custom, declClass = VarHandleStaticBaseComputer.class) //
     Object staticBase;
-    @Alias @RecomputeFieldValue(kind = RecomputeFieldValue.Kind.Custom, declClass = VarHandleFieldOffsetComputer.class) //
+    @Alias @RecomputeFieldValue(kind = RecomputeFieldValue.Kind.Custom, declClass = VarHandleFieldOffsetAsLongComputer.class) //
     long staticOffset;
-}
-
-class StaticAccessorFieldStaticBaseComputer implements FieldValueTransformerWithAvailability {
-    @Override
-    public FieldValueTransformerWithAvailability.ValueAvailability valueAvailability() {
-        return FieldValueTransformerWithAvailability.ValueAvailability.AfterAnalysis;
-    }
-
-    @Override
-    public Object transform(Object receiver, Object originalValue) {
-        Field field = ImageSingletons.lookup(VarHandleFeature.class).findVarHandleField(receiver);
-        if (field.getType().isPrimitive()) {
-            return StaticFieldsSupport.getStaticPrimitiveFields();
-        }
-        return StaticFieldsSupport.getStaticObjectFields();
-    }
 }